requirements.txt
# Telegram Bot Framework
aiogram==3.13.1

# AI/ML Libraries
openai==1.54.3
gigachat==0.1.33

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Redis
redis==5.2.0

# Web Framework (для админ-панели)
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Google Sheets (для базы знаний)
gspread==6.1.3
oauth2client==4.1.3

# Encryption (для ФЗ-152)
cryptography==43.0.3

# Logging
loguru==0.7.2

# HTTP requests
httpx==0.27.2
aiohttp==3.10.10

# Data processing
pandas==2.2.3
openpyxl==3.1.5

# Date/Time
python-dateutil==2.9.0.post0

# Environment
environs==11.0.0

